Migrated over Text, Span, Anchor and LinkPreview components
